backend/requirements.txt
django==6.0.3
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
google-genai
Pillow==12.1.1
python-dotenv==1.2.2
